requirements.txt
flask[async]==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0 